change bw from int to long. fix bug related to vertex ID in bigraph, need to take all vertices as ground set of possubset, not just left or right. start to implement degree heuristic.

+++ java/src/boolwidth/BooleanDecomposition.java
@@ -13,19 +13,19 @@ public class BooleanDecomposition {
 	 */
 	public final static String BOOLWIDTH_TIME_FIELD = "boolwidthTime";
 
-	public static int getBoolWidth(IDecomposition<?, ?, ?> dc) {
-		return (Integer) dc.getAttr(BOOLWIDTH_FIELD);
+	public static long getBoolWidth(IDecomposition<?, ?, ?> dc) {
+		return (Long) dc.getAttr(BOOLWIDTH_FIELD);
 	}
 
 	public static int getBoolWidthTime(IDecomposition<?, ?, ?> dc) {
 		return (Integer) dc.getAttr(BOOLWIDTH_TIME_FIELD);
 	}
 
-	public static void setBoolWidth(IDecomposition<?, ?, ?> dc, int bw) {
+	public static void setBoolWidth(IDecomposition<?, ?, ?> dc, long bw) {
 		dc.setAttr(BOOLWIDTH_FIELD, bw);
 	}
 
-	public static void setBoolWidthTime(IDecomposition<?, ?, ?> dc, int bw) {
+	public static void setBoolWidthTime(IDecomposition<?, ?, ?> dc, long bw) {
 		dc.setAttr(BOOLWIDTH_TIME_FIELD, bw);
 	}
 }
